Added more data for testing

+++ Search_Localize/src/ca/mcgill/ecse211/ultrasonic/UltrasonicController.java
@@ -2,7 +2,7 @@ package ca.mcgill.ecse211.ultrasonic;
 
 public interface UltrasonicController {
 
-  public void processUSData(int distance);
+  public void processUSData(double distance);
 
-  public int readUSDistance();
+  public double readUSDistance();
 }
